Refactor the snap support out into a new plugin
With much help from Simon Johnsson <simon.johnsson@canonical.com>, many thanks.
