+++ data/tests/fwupd.sh
@@ -50,6 +50,7 @@ run_test mtd-self-test
 run_test nitrokey-self-test
 run_test nvme-self-test
 run_test redfish-self-test
+run_test snap-self-test
 run_test synaptics-prometheus-self-test
 run_test tpm-self-test
 run_test uefi-dbx-self-test
