+++ libfwupdplugin/fu-common.h
@@ -138,6 +138,3 @@ void
 fu_xmlb_builder_insert_kx(XbBuilderNode *bn, const gchar *key, guint64 value) G_GNUC_NON_NULL(1);
 void
 fu_xmlb_builder_insert_kb(XbBuilderNode *bn, const gchar *key, gboolean value) G_GNUC_NON_NULL(1);
-
-gboolean
-fu_snap_is_in_snap(void);
