+++ plugins/snapd-uefi/tests/snapd.py
