+++ plugins/uefi-dbx/fu-uefi-dbx-device.c
@@ -8,40 +8,14 @@
 
 #include "fu-uefi-dbx-common.h"
 #include "fu-uefi-dbx-device.h"
-#include "fu-uefi-dbx-snapd-notifier.h"
 
 struct _FuUefiDbxDevice {
 	FuUefiDevice parent_instance;
-	FuUefiDbxSnapdNotifier *snapd_notifier;
 };
 
 G_DEFINE_TYPE(FuUefiDbxDevice, fu_uefi_dbx_device, FU_TYPE_UEFI_DEVICE)
 
 #define FU_UEFI_DBX_DEVICE_DEFAULT_REQUIRED_FREE (30 * 1024) /* bytes */
-
-void
-fu_uefi_dbx_device_set_snapd_notifier(FuUefiDbxDevice *self, FuUefiDbxSnapdNotifier *obs)
-{
-	g_set_object(&self->snapd_notifier, obs);
-}
-
-static gboolean
-fu_uefi_dbx_device_maybe_notify_snapd_prepare(FuUefiDbxDevice *self, GBytes *data, GError **error)
-{
-	if (self->snapd_notifier == NULL)
-		return TRUE;
-
-	return fu_uefi_dbx_snapd_notifier_dbx_update_prepare(self->snapd_notifier, data, error);
-}
-
-static gboolean
-fu_uefi_dbx_device_maybe_notify_snapd_cleanup(FuUefiDbxDevice *self, GError **error)
-{
-	if (self->snapd_notifier == NULL)
-		return TRUE;
-
-	return fu_uefi_dbx_snapd_notifier_dbx_update_cleanup(self->snapd_notifier, error);
-}
 
 static gboolean
 fu_uefi_dbx_device_write_firmware(FuDevice *device,
@@ -55,9 +29,6 @@ fu_uefi_dbx_device_write_firmware(FuDevice *device,
 	/* get default image */
 	fw = fu_firmware_get_bytes(firmware, error);
 	if (fw == NULL)
-		return FALSE;
-
-	if (!fu_uefi_dbx_device_maybe_notify_snapd_prepare(FU_UEFI_DBX_DEVICE(device), fw, error))
 		return FALSE;
 
 	/* write entire chunk to efivarsfs */
@@ -268,20 +239,6 @@ fu_uefi_dbx_device_set_progress(FuDevice *device, FuProgress *progress)
 	fu_progress_add_step(progress, FWUPD_STATUS_DEVICE_BUSY, 0, "reload");
 }
 
-static gboolean
-fu_uefi_dbx_device_cleanup(FuDevice *device,
-			   FuProgress *progress,
-			   FwupdInstallFlags flags,
-			   GError **error)
-{
-	FuUefiDbxDevice *self = FU_UEFI_DBX_DEVICE(device);
-
-	if (!fu_uefi_dbx_device_maybe_notify_snapd_cleanup(self, error))
-		return FALSE;
-
-	return TRUE;
-}
-
 static gchar *
 fu_uefi_dbx_device_convert_version(FuDevice *device, guint64 version_raw)
 {
@@ -322,28 +279,13 @@ fu_uefi_dbx_device_init(FuUefiDbxDevice *self)
 }
 
 static void
-fu_uefi_dbx_device_finalize(GObject *object)
-{
-	FuUefiDbxDevice *self = FU_UEFI_DBX_DEVICE(object);
-
-	if (self->snapd_notifier != NULL)
-		g_object_unref(self->snapd_notifier);
-
-	G_OBJECT_CLASS(fu_uefi_dbx_device_parent_class)->finalize(object);
-}
-
-static void
 fu_uefi_dbx_device_class_init(FuUefiDbxDeviceClass *klass)
 {
-	GObjectClass *object_class = G_OBJECT_CLASS(klass);
 	FuDeviceClass *device_class = FU_DEVICE_CLASS(klass);
 	device_class->probe = fu_uefi_dbx_device_probe;
 	device_class->reload = fu_uefi_dbx_device_reload;
 	device_class->write_firmware = fu_uefi_dbx_device_write_firmware;
 	device_class->prepare_firmware = fu_uefi_dbx_device_prepare_firmware;
 	device_class->set_progress = fu_uefi_dbx_device_set_progress;
-	device_class->cleanup = fu_uefi_dbx_device_cleanup;
 	device_class->convert_version = fu_uefi_dbx_device_convert_version;
-
-	object_class->finalize = fu_uefi_dbx_device_finalize;
 }
