+++ plugins/uefi-dbx/fu-uefi-dbx-device.h
@@ -8,10 +8,5 @@
 
 #include <fwupdplugin.h>
 
-#include "fu-uefi-dbx-snapd-notifier.h"
-
 #define FU_TYPE_UEFI_DBX_DEVICE (fu_uefi_dbx_device_get_type())
 G_DECLARE_FINAL_TYPE(FuUefiDbxDevice, fu_uefi_dbx_device, FU, UEFI_DBX_DEVICE, FuUefiDevice)
-
-void
-fu_uefi_dbx_device_set_snapd_notifier(FuUefiDbxDevice *self, FuUefiDbxSnapdNotifier *obs);
